[YS-224] refactor: 회원가입 연구자 디렉토리 구조 변경
