+++ src/app/join/components/JoinCheckboxContainer/JoinCheckbox/JoinCheckbox.styles.ts
