+++ src/app/join/components/Participant/JoinEmailStep/JoinEmailStep.tsx
@@ -1,12 +1,12 @@
 import { useFormContext, useWatch } from 'react-hook-form';
 
 import { nextButton } from './JoinEmailStep.styles';
-import JoinCheckboxContainer from '../../JoinEmailStep/JoinCheckboxContainer/JoinCheckboxContainer';
 import JoinInput from '../../JoinInput/JoinInput';
 
 import useServiceAgreeCheck from '@/app/join/hooks/useServiceAgreeCheck';
 import { joinContentContainer, joinForm } from '@/app/join/JoinPage.styles';
 import { ParticipantJoinSchemaType } from '@/schema/join/ParticipantJoinSchema';
+import JoinCheckboxContainer from '../../JoinCheckboxContainer/JoinCheckboxContainer';
 
 interface JoinEmailStepProps {
   onNext: () => void;
