+++ src/app/join/components/Researcher/JoinInfoStep/JoinInfoStep.styles.ts
