+++ src/app/join/components/Researcher/JoinInfoStep/JoinInfoStep.tsx
@@ -1,8 +1,8 @@
 import { useFormContext, useWatch } from 'react-hook-form';
 
 import { joinButton } from './JoinInfoStep.styles';
-import { joinContentContainer, joinForm } from '../../JoinPage.styles';
-import JoinInput from '../JoinInput/JoinInput';
+import { joinContentContainer, joinForm } from '../../../JoinPage.styles';
+import JoinInput from '../../JoinInput/JoinInput';
 
 import { ResearcherJoinSchemaType } from '@/schema/join/ResearcherJoinSchema';
 
@@ -10,7 +10,6 @@ interface JoinInfoStepProps {
   handleSubmit: () => void;
 }
 
-// TODO: blur 될 때 trigger + 입력값 초기화 안되도록 개선 필요
 const JoinInfoStep = ({ handleSubmit }: JoinInfoStepProps) => {
   const { control } = useFormContext<ResearcherJoinSchemaType>();
   const values = useWatch({ name: ['name', 'univName', 'major'], control });
