+++ src/app/join/components/Researcher/index.tsx
@@ -1,5 +1,5 @@
-import JoinEmailStep from '../JoinEmailStep/JoinEmailStep';
-import JoinInfoStep from '../JoinInfoStep/JoinInfoStep';
+import JoinEmailStep from './JoinEmailStep/JoinEmailStep';
+import JoinInfoStep from './JoinInfoStep/JoinInfoStep';
 
 export const Researcher = {
   EmailStep: JoinEmailStep,
